requirements.txt
langgraph
openai
Pillow
